some functions with calendars and events

+++ app/Http/Controllers/CalendarController.php
@@ -4,9 +4,62 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Calendar;
+use App\Models\User_calendar;
+use App\Models\User;
 
 class CalendarController extends Controller
 {
+
+     /**
+     * Display the specified resource.
+     * @param  \App\Models\Calendar  $calendar
+     * @return \Illuminate\Http\Response
+     */
+    public function createCalendar(Request $request)
+    {
+
+        if(!$this->check_auth($request)) {
+            return response("You have no auth", 401);
+        }
+
+        $validated = $request->validate([
+            'title' => 'required|string',
+            'description' => 'string',
+        ]);
+        $user_id = $this->get_auth($request)->id;
+
+        $validated["user_id"] = $user_id;
+
+        $calendar = Calendar::create($validated);
+
+        User_calendar::create([
+            'user_id' => $user_id,
+            'calendar_id' => $calendar->id,
+        ]);
+
+        return $calendar;
+
+    }
+
+    /**
+     * Display a listing of the resource.
+     * @return \Illuminate\Http\Response
+     */
+    public function getUsersForCalendar(Request $request, $id)
+    {
+        if(!$this->check_auth($request)) {
+            return response("You have no auth", 403);
+        }
+
+        $users = \DB::table("user_calendars")->where("calendar_id", $id)->get();
+
+        foreach($users as $user) {
+            $user->users_data = User::find($user->user_id);
+        }
+
+        return $users;
+    }
+
     /**
      * Display a listing of the resource.
      * @return \Illuminate\Http\Response
@@ -14,10 +67,16 @@ class CalendarController extends Controller
     public function getCalendarsForUser(Request $request)
     {
         if(!$this->check_auth($request)) {
-            return response("Forbidden", 403);
+            return response("You have no auth", 403);
         }
 
-        return \DB::select("select * from calendars where user_id = ".$this->get_auth($request)->id);
+        $calendars = \DB::table("user_calendars")->where("user_id", $this->get_auth($request)->id)->get();
+
+        foreach($calendars as $calendar) {
+            $calendar->calendar_data = Calendar::find($calendar->calendar_id);
+        }
+
+        return $calendars;
     }
 
     /**
@@ -35,14 +94,23 @@ class CalendarController extends Controller
             return response("You have no auth", 401);
         }
 
-        if($this->get_auth($request)->id == $calendar->user_id) {
+        $calendar = \DB::table("calendars")->where("id", $id)->first();
+
+        if(!$calendar){
+            return response("Calendar not found", 404);
+        }
+
+        $calendar->users = \DB::table("user_calendars")->where("calendar_id", $id)->pluck("user_id")->toArray();
+
+        if(in_array($this->get_auth($request)->id, $calendar->users)) {
             return $calendar;
         }
 
         return response("Something wring", 403);
 
     }
-        /**
+
+    /**
      * Update the specified resource in storage.
      * @param  \Illuminate\Http\Request  $request
      * @param  \App\Models\Calendar  $calendar
@@ -89,8 +157,10 @@ class CalendarController extends Controller
             return response("You have no auth", 401);
         }
 
-        if($this->get_auth($request)->id == $calendar->user_id) {
+        if($this->get_auth($request)->id == $calendar->user_id && $calendar->main == 0) {
             return Calendar::destroy($id);
         }
+
+        return response("Something wring", 403);
     }
 }
